Set output filename in makefile instead of from command line
This was causing a major bug when building the peripheral drivers where output library name was being overridden by setting the PROJECT variable globally.  Need to figure out a way to work around this, because at the moment this is a clunky way to match the output file name to the workspace folder name for the launch configuration file.  I want to be able to set this with VSCode's integrated workspace folder name variable...